OAHashTable: Added tableLength field with (protected) getter, which holds the length of the table array ("m").
Implemented LPHashTable, and DoubleHashTable excluding the choice of
the stepHash. baseHash is the base hash function used the determine the
first index in the probing sequence, which is the same for all the hash
table implementations, and stepHash is the hash used to determine the
step of the probing sequence in DoubleHashTable.